refactor: move button bar logic to separate ProjectPaperView component
This again removes logic from the PaperView and makes it easier to manage.
The whole review logic is now in ProjectPaperView.
